+++ src/Symfony/Component/Translation/Tests/TranslatorTest.php
@@ -391,48 +391,6 @@ class TranslatorTest extends TestCase
         $this->assertEquals($expected, $translator->trans($id, [], '', 'fr'));
     }
 
-    /**
-     * @dataProvider getTransChoiceTests
-     * @group legacy
-     */
-    public function testTransChoice($expected, $id, $translation, $number, $parameters, $locale, $domain)
-    {
-        $translator = new Translator('en');
-        $translator->addLoader('array', new ArrayLoader());
-        $translator->addResource('array', [(string) $id => $translation], $locale, $domain);
-
-        $this->assertEquals($expected, $translator->transChoice($id, $number, $parameters, $domain, $locale));
-    }
-
-    /**
-     * @dataProvider      getInvalidLocalesTests
-     * @expectedException \Symfony\Component\Translation\Exception\InvalidArgumentException
-     * @group legacy
-     */
-    public function testTransChoiceInvalidLocale($locale)
-    {
-        $translator = new Translator('en');
-        $translator->addLoader('array', new ArrayLoader());
-        $translator->addResource('array', ['foo' => 'foofoo'], 'en');
-
-        $translator->transChoice('foo', 1, [], '', $locale);
-    }
-
-    /**
-     * @dataProvider      getValidLocalesTests
-     * @group legacy
-     */
-    public function testTransChoiceValidLocale($locale)
-    {
-        $translator = new Translator('en');
-        $translator->addLoader('array', new ArrayLoader());
-        $translator->addResource('array', ['foo' => 'foofoo'], 'en');
-
-        $translator->transChoice('foo', 1, [], '', $locale);
-        // no assertion. this method just asserts that no exception is thrown
-        $this->addToAssertionCount(1);
-    }
-
     public function getTransFileTests()
     {
         return [
@@ -477,32 +435,6 @@ class TranslatorTest extends TestCase
             ['Symfony est super!', $messages, 'symfony.is.great'],
             ['Foo Bar Baz', $messages, 'foo.bar.baz'],
             ['Foo Baz', $messages, 'foo.baz'],
-        ];
-    }
-
-    public function getTransChoiceTests()
-    {
-        return [
-            ['Il y a 0 pomme', '{0} There are no appless|{1} There is one apple|]1,Inf] There is %count% apples', '[0,1] Il y a %count% pomme|]1,Inf] Il y a %count% pommes', 0, [], 'fr', ''],
-            ['Il y a 1 pomme', '{0} There are no appless|{1} There is one apple|]1,Inf] There is %count% apples', '[0,1] Il y a %count% pomme|]1,Inf] Il y a %count% pommes', 1, [], 'fr', ''],
-            ['Il y a 10 pommes', '{0} There are no appless|{1} There is one apple|]1,Inf] There is %count% apples', '[0,1] Il y a %count% pomme|]1,Inf] Il y a %count% pommes', 10, [], 'fr', ''],
-
-            ['Il y a 0 pomme', 'There is one apple|There is %count% apples', 'Il y a %count% pomme|Il y a %count% pommes', 0, [], 'fr', ''],
-            ['Il y a 1 pomme', 'There is one apple|There is %count% apples', 'Il y a %count% pomme|Il y a %count% pommes', 1, [], 'fr', ''],
-            ['Il y a 10 pommes', 'There is one apple|There is %count% apples', 'Il y a %count% pomme|Il y a %count% pommes', 10, [], 'fr', ''],
-
-            ['Il y a 0 pomme', 'one: There is one apple|more: There is %count% apples', 'one: Il y a %count% pomme|more: Il y a %count% pommes', 0, [], 'fr', ''],
-            ['Il y a 1 pomme', 'one: There is one apple|more: There is %count% apples', 'one: Il y a %count% pomme|more: Il y a %count% pommes', 1, [], 'fr', ''],
-            ['Il y a 10 pommes', 'one: There is one apple|more: There is %count% apples', 'one: Il y a %count% pomme|more: Il y a %count% pommes', 10, [], 'fr', ''],
-
-            ['Il n\'y a aucune pomme', '{0} There are no apples|one: There is one apple|more: There is %count% apples', '{0} Il n\'y a aucune pomme|one: Il y a %count% pomme|more: Il y a %count% pommes', 0, [], 'fr', ''],
-            ['Il y a 1 pomme', '{0} There are no apples|one: There is one apple|more: There is %count% apples', '{0} Il n\'y a aucune pomme|one: Il y a %count% pomme|more: Il y a %count% pommes', 1, [], 'fr', ''],
-            ['Il y a 10 pommes', '{0} There are no apples|one: There is one apple|more: There is %count% apples', '{0} Il n\'y a aucune pomme|one: Il y a %count% pomme|more: Il y a %count% pommes', 10, [], 'fr', ''],
-
-            ['Il y a 0 pomme', new StringClass('{0} There are no appless|{1} There is one apple|]1,Inf] There is %count% apples'), '[0,1] Il y a %count% pomme|]1,Inf] Il y a %count% pommes', 0, [], 'fr', ''],
-
-            // Override %count% with a custom value
-            ['Il y a quelques pommes', 'one: There is one apple|more: There are %count% apples', 'one: Il y a %count% pomme|more: Il y a quelques pommes', 2, ['%count%' => 'quelques'], 'fr', ''],
         ];
     }
 
@@ -553,46 +485,6 @@ class TranslatorTest extends TestCase
 
         $translator->addResource('array', ['some_message' => 'Hi {name}'], 'en', 'messages+intl-icu');
         $this->assertSame('Hi Bob', $translator->trans('some_message', ['%name%' => 'Bob']));
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testTransChoiceFallback()
-    {
-        $translator = new Translator('ru');
-        $translator->setFallbackLocales(['en']);
-        $translator->addLoader('array', new ArrayLoader());
-        $translator->addResource('array', ['some_message2' => 'one thing|%count% things'], 'en');
-
-        $this->assertEquals('10 things', $translator->transChoice('some_message2', 10, ['%count%' => 10]));
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testTransChoiceFallbackBis()
-    {
-        $translator = new Translator('ru');
-        $translator->setFallbackLocales(['en_US', 'en']);
-        $translator->addLoader('array', new ArrayLoader());
-        $translator->addResource('array', ['some_message2' => 'one thing|%count% things'], 'en_US');
-
-        $this->assertEquals('10 things', $translator->transChoice('some_message2', 10, ['%count%' => 10]));
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testTransChoiceFallbackWithNoTranslation()
-    {
-        $translator = new Translator('ru');
-        $translator->setFallbackLocales(['en']);
-        $translator->addLoader('array', new ArrayLoader());
-
-        // consistent behavior with Translator::trans(), which returns the string
-        // unchanged if it can't be found
-        $this->assertEquals('some_message2', $translator->transChoice('some_message2', 10, ['%count%' => 10]));
     }
 }